aggregator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
anthropic==0.28.0
openai==1.30.0
boto3==1.34.0
azure-storage-blob==12.19.0
requests==2.32.3
pyyaml==6.0.1
